+++ xenon-common/src/main/java/com/vmware/xenon/common/ServiceHost.java
@@ -2275,7 +2275,7 @@ public class ServiceHost {
         }
 
         ServiceDocumentDescription documentDescription = buildDocumentDescription(service);
-        QueryFilter queryFilter = ctx.getResourceQueryFilter();
+        QueryFilter queryFilter = ctx.getResourceQueryFilter(op.getAction());
         if (queryFilter == null || !queryFilter.evaluate(document, documentDescription)) {
             return false;
         }